Safari-611.1.10/LayoutTests/imported/w3c/web-platform-tests/html/webappapis/scripting/events/invalid-uncompiled-raw-handler-compiled-late-expected.txt

CONSOLE MESSAGE: SyntaxError: Parser error
CONSOLE MESSAGE: SyntaxError: Parser error

FAIL Invalid uncompiled raw handlers should only be compiled when about to call them. assert_array_equals: lengths differ, expected 3 got 4
